added background image to about us page
