+++ upin/src/app/about-us/page.tsx
@@ -1,0 +1,58 @@
+import React from 'react'
+import SouthOutlinedIcon from '@mui/icons-material/SouthOutlined';
+import type { Metadata } from "next";
+import bgImg from "@/../public/Screen Shot 2020-03-12 at 9.26.39 AM.png"
+import Image from 'next/image';
+
+export const metadata: Metadata = {
+  title: "About Us | Upin",
+  description: "Generated by create next app",
+};
+
+const aboutUs = () => {
+  return (
+    <div className='relative w-full min-h-screen'>
+      <Image 
+        src={bgImg}
+        alt='Image of background'
+        layout='fill'
+        objectFit='cover'
+        className=' mix-blend-overlay opacity-75'
+      />
+      <div className='relative z-10 flex flex-col items-center text-black'>
+        <div className='flex justify-center pt-16'>
+          <h1 className='font-serif text-6xl mr-3 font-extrabold text-black'>About Us</h1>
+        </div>
+        <div className='flex justify-center mt-5 mb-8'>
+          <p className='font-serif text-3xl text-center px-4 font-bold'>We believe we are one and we value the importance of connecting with people in real life.</p>
+        </div>
+        <div className='flex justify-center px-8'>
+          <div className='border border-black rounded-md bg-green-500 bg-opacity-75 shadow-2xl shadow-gray-70 p-6'>
+            <p className='font-serif text-2xl'>
+              <span className='font-bold'>Our Mission - </span>
+              Upin is 100% fully committed to creating the best platform for keeping your real social life in one place. We create the best experience for hosting and joining local activities by making it more convenient, organized, and safe. Our mission is to create more real-life interaction by being able to create & join gatherings of people anywhere at any time. We are focused on creating more peace and oneness amongst each other by being more openly social. With Upin, you will never feel alone again.
+            </p>
+          </div>
+        </div>
+        <div className='flex justify-center m-5'>
+          <SouthOutlinedIcon className='w-20 h-20'/>
+        </div>
+        <div className='flex justify-center px-8'>
+          <div className='border border-black rounded-md bg-green-500 bg-opacity-75 shadow-2xl shadow-gray-700 p-6'>
+            <p className='font-serif text-2xl'><span className='font-bold'>We Value & Provide -</span> Connection, Relationships, Community. Connection is key to growing in a community. It takes strong relationships to have strong community, and with strong community we grow together in power!</p>
+          </div>
+        </div>
+        <div className='flex justify-center m-5'>
+          <SouthOutlinedIcon className='w-20 h-20'/>
+        </div>
+        <div className='flex justify-center px-8'>
+          <div className='border border-black rounded-md bg-green-500 bg-opacity-75 shadow-2xl shadow-gray-700 p-6 mb-5'>
+            <p className='font-serif text-2xl'><span className='font-bold'>Our Philosophy -</span> stands behind living longer & stronger with more loving community & social interaction. We are backed by our philosophy, mission, & life guidelines at <span className='underline'>All Is One Movement.</span></p>
+          </div>
+        </div>
+      </div>
+    </div>
+  )
+}
+
+export default aboutUs
