+++ upin/src/app/components/EventMarker.tsx
@@ -10,7 +10,7 @@ interface MarkerProps {
 const EventMarker: React.FC<MarkerProps> = ({ lat, lng, title, description }) => {
   return (
     <div>
-      {/* Customize the appearance of your marker here */}
+      
       <h3>{title}</h3>
       <p>{description}</p>
     </div>
